Add a shared CLI test harness and migrate cli_* tests to it
Add `tests/common/cli.rs` with `run` (invoke the ryl binary, return code+stdout+stderr)
and `command_output`, and migrate the 37 `cli_*` integration tests that each re-defined
those helpers to `mod common; use common::cli::...` instead. Behaviour is unchanged; the
full suite passes.

Part of #277 (item 2).
