+++ tests/cli_comments_rule.rs
@@ -3,13 +3,8 @@ use std::process::Command;
 
 use tempfile::tempdir;
 
-fn run(cmd: &mut Command) -> (i32, String, String) {
-    let out = cmd.output().expect("process");
-    let code = out.status.code().unwrap_or(-1);
-    let stdout = String::from_utf8_lossy(&out.stdout).into_owned();
-    let stderr = String::from_utf8_lossy(&out.stderr).into_owned();
-    (code, stdout, stderr)
-}
+mod common;
+use common::cli::run;
 
 #[test]
 fn comments_rule_emits_diagnostics() {
