+++ tests/cli_fix_ignore_new_lines.rs
@@ -2,13 +2,8 @@ use std::fs;
 use std::process::Command;
 use tempfile::tempdir;
 
-fn run(cmd: &mut Command) -> (i32, String, String) {
-    let out = cmd.output().expect("process");
-    let code = out.status.code().unwrap_or(-1);
-    let stdout = String::from_utf8_lossy(&out.stdout).into_owned();
-    let stderr = String::from_utf8_lossy(&out.stderr).into_owned();
-    (code, stdout, stderr)
-}
+mod common;
+use common::cli::run;
 
 #[test]
 fn fix_respects_new_lines_ignore_for_eof_newline() {
