+++ tests/common/mod.rs
@@ -1,1 +1,2 @@
+pub mod cli;
 pub mod fake_env;
