+++ tests/flow_collection_cli.rs
@@ -13,13 +13,8 @@ struct CliSuite {
     empty_message: &'static str,
 }
 
-fn run(cmd: &mut Command) -> (i32, String, String) {
-    let out = cmd.output().expect("process");
-    let code = out.status.code().unwrap_or(-1);
-    let stdout = String::from_utf8_lossy(&out.stdout).into_owned();
-    let stderr = String::from_utf8_lossy(&out.stderr).into_owned();
-    (code, stdout, stderr)
-}
+mod common;
+use common::cli::run;
 
 fn assert_command_output(
     cmd: &mut Command,
